Rename TransformerLayerState to LayerState; move MTP callables to common
Three follow-up cleanups on the schedule-plan plumbing:

* Rename the per-layer placeholder class TransformerLayerState to LayerState
  in core/models/common/utils.py (the name lives in common now and the
  Transformer prefix was misleading).

* Move build_mtp_layer_callables and the build_layer_callables dispatcher
  out of core/models/gpt/fine_grained_callables.py and into the new
  core/models/common/fine_grained_callables.py. MTP is shared between
  GPTModel and HybridModel; the dispatcher's job is to dispatch on layer
  type and is naturally common too. Only build_transformer_layer_callables
  stays in gpt/ since it depends on GPT's MoE wiring.

* Drop the gpt/ re-export block (the __all__ list and the bulk import from
  common.utils). Callers now import the moved names directly from
  core/models/common/utils.py and core/models/common/fine_grained_callables
  .py. Gpt/fine_grained_callables.py is GPT-only again.

Restore the explanatory comments inside HybridModel._preprocess and
_postprocess that the earlier forward-refactor lost — they were originally
inline in forward() and got dropped when the body moved into the helper
methods. The new comments explain decoder-input handling, rotary cos/sin
discard, sequence-parallel gather rationale, and the speculative-decoding
ordering constraint.
