+++ megatron/core/models/common/model_chunk_schedule_plan.py
@@ -70,10 +70,10 @@ class TransformerLayerSchedulePlan:
         The event and chunk_state are binded to the TransformerModelChunkSchedulePlan
         and shared across all layers in the model chunk.
         """
-        from megatron.core.models.common.utils import TransformerLayerState
+        from megatron.core.models.common.utils import LayerState
 
         self.config = layer.config
-        self.layer_state = TransformerLayerState()
+        self.layer_state = LayerState()
         self.chunk_state = chunk_state
         self.layer = layer
         self.event = event
@@ -111,8 +111,8 @@ class TransformerLayerSchedulePlan:
         Builds the callable nodes for the transformer/mtp layer:
             attn, mlp, moe_dispatch and moe_combine, and mtp_post_process.
         """
+        from megatron.core.models.common.fine_grained_callables import build_layer_callables
         from megatron.core.models.common.utils import TransformerLayerNode
-        from megatron.core.models.gpt.fine_grained_callables import build_layer_callables
         from megatron.core.transformer.moe.moe_layer import MoELayer
         from megatron.core.transformer.multi_token_prediction import MultiTokenPredictionLayer
